superadmin is_approved api

+++ controllers/organizer/events/remove.js
@@ -10,7 +10,7 @@ exports.removeevent = async (req, res) => {
         const { eventid } = req.body;
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
         let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
-        if(organizerData && organizerData.status == true && organizerData.mobileverified == true){
+        if(organizerData && organizerData.status == true && organizerData.mobileverified == true && organizerData.is_approved == true){
             if (eventid && eventid != '' && mongoose.Types.ObjectId.isValid(eventid)) {
                 await primary.model(constants.MODELS.events, eventModel).findByIdAndRemove(eventid);
                 return responseManager.onSuccess('Organizer event removed successfully!', 1, res);
